requirements.txt
Flask==2.2.2
opencv-python-headless==4.7.0.72
mediapipe==0.10.18
tensorflow==2.18.0
numpy>=1.26.0,<2.1.0
requests==2.28.1
gunicorn==20.1.0
Werkzeug==2.2.3
azure-storage-queue
